Remove int casting from shader class

+++ VoxelEngine/src/core/shader.cpp
@@ -87,5 +87,5 @@ void Shader::setInt(const std::string& name, int value) {
 };
 
 void Shader::setFloat(const std::string& name, float value) {
-	glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+	glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
 };
